- Refactor generator structure - Make generator path independent - Add perl critic - Improve gulp watch performance - Add watcher for file delete
