--- app/prompts/user.js
@@ -1,8 +1,0 @@
-module.exports = function (config) {
-    return {
-        type: 'input',
-        name: 'vm-usr',
-        message: 'VM Username',
-        default: config || 'root'
-    }
-};
