+++ install/cli-tools.js
