+++ install/prompts/domain.js
