+++ install/prompts/password.js
@@ -1,0 +1,10 @@
+'use strict';
+
+module.exports = function (config) {
+    return {
+        type: 'password',
+        name: 'vm-pwd',
+        message: 'VM Password',
+        default: config || 'qwert6'
+    }
+};
