+++ install/prompts/store.js
