+++ install/prompts/user.js
@@ -3,8 +3,8 @@
 module.exports = function (config) {
     return {
         type: 'input',
-        name: 'vm-pwd',
-        message: 'VM Password',
-        default: config || 'qwert6'
+        name: 'vm-usr',
+        message: 'VM Username',
+        default: config || 'root'
     }
 };
