+++ install/prompts/watch.js
