+++ install/templates/dotfiles/.editorconfig
