+++ install/templates/dotfiles/.eslintrc
