+++ install/templates/gulp/lib/file-watch.js
@@ -1,0 +1,48 @@
+/*eslint-env node*/
+'use strict';
+
+var _ = require('lodash'),
+    gulp = require('gulp'),
+    GulpSSH = require('gulp-ssh'),
+    path = require('path'),
+    config = require('../config');
+
+var buildPath = [config.webroot, 'StoreTypes', config.version, config.store].join('/'),
+
+    sshConfig = {
+        host: config['vm-domain'],
+        port: 22,
+        username: config['vm-usr'],
+        password: config['vm-pwd']
+    },
+
+    ssh = new GulpSSH({
+        ignoreErrors: true,
+        sshConfig: sshConfig
+    });
+
+module.exports = function (pattern, onChange, onDelete) {
+    if (typeof pattern === 'string') {
+        pattern = pattern.split('my-fake-split');
+    }
+
+    pattern = _.map(pattern, function (element) {
+        return config['cartridges-local'] + element;
+    });
+
+    return gulp.watch(pattern, {interval: 1500}, function (file) {
+        var dest;
+
+        file.path.replace(path.sep, '/');
+
+        if (file.type === 'deleted' && onDelete !== undefined) {
+            dest = buildPath + file.path.replace(/^.*?\/Data\/Public/, '');
+            return onDelete(file.path, ssh.shell('rm ' + dest));
+        }
+
+        if (file.type === 'changed' && onChange !== undefined) {
+            dest = buildPath + path.dirname(file.path.replace(/^.*?\/Data\/Public/, ''));
+            return onChange(file.path, dest, ssh.dest(dest));
+        }
+    });
+};
