+++ install/templates/gulp/package.json
