+++ package.json
@@ -1,7 +1,8 @@
 {
   "name": "generator-epages6",
-  "version": "0.1.0",
+  "version": "0.2.0",
   "description": "",
+  "preferGlobal": true,
   "files": [
     "app",
     "router"
